+++ CHANGELOG.md
@@ -7,6 +7,7 @@ Changelog
 
 - Add Bubble component [PR 31](https://github.com/input-output-hk/react-polymorph/pull/31)
 - Use Bubble component in Options component [PR 31](https://github.com/input-output-hk/react-polymorph/pull/31)
+- Add Tooltip component [PR 32](https://github.com/input-output-hk/react-polymorph/pull/32)
 
 ### Fixes